Use journalctl for test logs

+++ tests/udp.sh
@@ -11,8 +11,6 @@ DEV=veth0
 DEV_A=${DEV}a
 DEV_B=${DEV}b
 
-LOG_TO_FILE=tests/udp.log
-
 # Setup network
 if [ ! -e /run/netns/$NS ]; then
   ip netns add $NS
@@ -34,7 +32,18 @@ ip netns exec $NS ip link set $DEV_B up
 
 # Start Gatekeeper
 systemctl reset-failed
-systemd-run --service-type=notify --same-dir --unit=gatekeeper-udp --setenv=LAN=$DEV_A --setenv=LOG_TO_FILE=$LOG_TO_FILE --quiet build/debug_gatekeeper
+systemd-run --service-type=notify --same-dir --unit=gatekeeper-udp --setenv=LAN=$DEV_A --quiet build/debug_gatekeeper
+GATEKEEPER_STATUS=$?
+INVOCATION_ID=$(systemctl show --value -p InvocationID gatekeeper-udp)
+
+if [ $GATEKEEPER_STATUS -ne 0 ]; then
+  echo "Gatekeeper failed to start. Status code: $GATEKEEPER_STATUS"
+  echo "Gatekeeper log:"
+  journalctl _SYSTEMD_INVOCATION_ID=$INVOCATION_ID
+  exit 1
+fi
+
+echo "Use 'journalctl _SYSTEMD_INVOCATION_ID=$INVOCATION_ID' to see Gatekeeper logs"
 
 # Start dhclient
 cp -f test-dhclient-hook /etc/dhcp/dhclient-enter-hooks.d/
